Split location list view into project-based list views
