+++ packages/fhir-location-management/src/helpers/utils.ts
@@ -116,12 +116,3 @@ export const useGetLocation = (baseUrl: string, locId?: string) => {
     enabled: !!locId,
   });
 };
-
-// TODO - move to fhir-helpers
-const locationValueSet = 'http://hl7.org/fhir/ValueSet/location-physical-type';
-enum PhysicalType {
-  BIULDING = 'bu',
-}
-export const eusmPhysicalLocationsFilterParams = {
-  type: `${locationValueSet}|${PhysicalType.BIULDING}`,
-};
